+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 ### 2.1.0
 
 - Add selection animation 
-- Mode customisation with `showCheckView` 
+- Mode customisation with `showCheckView` and `performTapAnimation`
 - Bump FBSDK to 4.28
 
 ### 2.0.2